refactor container dashboard code to mixin
