+++ app/services/container_dashboard_service.rb
@@ -1,11 +1,12 @@
 class ContainerDashboardService
   include UiServiceMixin
-  CPU_USAGE_PRECISION = 2 # 2 decimal points
-  REALTIME_TIME_RANGE = 10 # 10 minutes
+  include ContainerServiceMixin
+
 
   def initialize(provider_id, controller)
     @provider_id = provider_id
     @ems = ManageIQ::Providers::ContainerManager.find(@provider_id) unless @provider_id.blank?
+    @resource = @ems || ManageIQ::Providers::ContainerManager.all
     @controller = controller
   end
 
@@ -176,191 +177,17 @@ class ContainerDashboardService
     hourly_heatmaps || realtime_heatmaps
   end
 
-  def fill_ems_utilization(m, time, used_cpu, used_mem, total_cpu, total_mem)
-    used_cpu[time] += m.v_derived_cpu_total_cores_used if m.v_derived_cpu_total_cores_used.present?
-    used_mem[time] += m.derived_memory_used if m.derived_memory_used.present?
-    total_cpu[time] += m.derived_vm_numvcpus if m.derived_vm_numvcpus.present?
-    total_mem[time] += m.derived_memory_available if m.derived_memory_available.present?
-  end
-
-  def realtime_ems_utilization
-    used_cpu = Hash.new(0)
-    used_mem = Hash.new(0)
-    total_cpu = Hash.new(0)
-    total_mem = Hash.new(0)
-
-    realtime_provider_metrics.each do |m|
-      minute = m.timestamp.beginning_of_minute.utc
-      fill_ems_utilization(m, minute, used_cpu, used_mem, total_cpu, total_mem)
-    end
-
-    {
-      :interval_name => "realtime",
-      :xy_data       => ems_utilization_data(used_cpu, total_cpu, used_mem, total_mem) || {:cpu => nil, :mem => nil}
-    }
-  end
-
-  def hourly_ems_utilization
-    used_cpu = Hash.new(0)
-    used_mem = Hash.new(0)
-    total_cpu = Hash.new(0)
-    total_mem = Hash.new(0)
-
-    hourly_provider_metrics.each do |m|
-      hour = m.timestamp.beginning_of_hour.utc
-      fill_ems_utilization(m, hour, used_cpu, used_mem, total_cpu, total_mem)
-    end
-
-    if used_cpu.any?
-      {
-        :interval_name => "hourly",
-        :xy_data       => ems_utilization_data(used_cpu, total_cpu, used_mem, total_mem)
-      }
-    end
-  end
-
-  def daily_ems_utilization
-    used_cpu = Hash.new(0)
-    used_mem = Hash.new(0)
-    total_cpu = Hash.new(0)
-    total_mem = Hash.new(0)
-
-    daily_provider_metrics.each do |metric|
-      date = metric.timestamp.strftime("%Y-%m-%d")
-      fill_ems_utilization(metric, date, used_cpu, used_mem, total_cpu, total_mem)
-    end
-
-    if used_cpu.any?
-      {
-        :interval_name => "daily",
-        :xy_data       => ems_utilization_data(used_cpu, total_cpu, used_mem, total_mem)
-      }
-    end
-  end
-
-  def ems_utilization_data(used_cpu, total_cpu, used_mem, total_mem)
-    if used_cpu.any?
-      {
-        :cpu => {
-          :used  => used_cpu.values.last.round,
-          :total => total_cpu.values.last.round,
-          :xData => used_cpu.keys,
-          :yData => used_cpu.values.map(&:round)
-        },
-        :mem => {
-          :used  => (used_mem.values.last / 1024.0).round,
-          :total => (total_mem.values.last / 1024.0).round,
-          :xData => used_mem.keys,
-          :yData => used_mem.values.map { |m| (m / 1024.0).round }
-        }
-      }
-    end
-  end
-
   def ems_utilization
-    daily_ems_utilization || hourly_ems_utilization || realtime_ems_utilization
-  end
-
-  def realtime_network_metrics
-    realtime_network_metrics = Hash.new(0)
-    realtime_provider_metrics.each do |m|
-      minute = m.timestamp.beginning_of_minute.utc
-      realtime_network_metrics[minute] += m.net_usage_rate_average if m.net_usage_rate_average.present?
-    end
-
-    {
-      :interval_name => "realtime",
-      :xy_data       => trend_data(realtime_network_metrics)
-    }
-  end
-
-  def hourly_network_metrics
-    hourly_network_metrics = Hash.new(0)
-    hourly_provider_metrics.each do |m|
-      hour = m.timestamp.beginning_of_hour.utc
-      hourly_network_metrics[hour] += m.net_usage_rate_average if m.net_usage_rate_average.present?
-    end
-
-    if hourly_network_metrics.size > 1
-      {
-        :interval_name => "hourly",
-        :xy_data       => trend_data(hourly_network_metrics)
-      }
-    end
-  end
-
-  def daily_network_metrics
-    daily_network_metrics = Hash.new(0)
-    daily_provider_metrics.each do |m|
-      day = m.timestamp.strftime("%Y-%m-%d")
-      daily_network_metrics[day] += m.net_usage_rate_average if m.net_usage_rate_average.present?
-    end
-
-    if daily_network_metrics.size > 1
-      {
-        :interval_name => "daily",
-        :xy_data       => trend_data(daily_network_metrics)
-      }
-    end
+    daily_utilization || hourly_utilization || realtime_utilization || empty_utilization_trend_data
   end
 
   def network_metrics
-    daily_network_metrics || hourly_network_metrics || realtime_network_metrics
-  end
-
-  def fill_pod_metrics(m, time, pod_create_trend, pod_delete_trend)
-    pod_create_trend[time] += m.stat_container_group_create_rate if m.stat_container_group_create_rate.present?
-    pod_delete_trend[time] += m.stat_container_group_delete_rate if m.stat_container_group_delete_rate.present?
-  end
-
-  def hourly_pod_metrics
-    hourly_pod_create_trend = Hash.new(0)
-    hourly_pod_delete_trend = Hash.new(0)
-    hourly_provider_metrics.each do |m|
-      hour = m.timestamp.beginning_of_hour.utc
-      fill_pod_metrics(m, hour, hourly_pod_create_trend, hourly_pod_delete_trend)
-    end
-
-    {
-      :interval_name => "hourly",
-      :xy_data       => create_delete_data(hourly_pod_create_trend, hourly_pod_delete_trend)
-    }
-  end
-
-  def daily_pod_metrics
-    daily_pod_create_trend = Hash.new(0)
-    daily_pod_delete_trend = Hash.new(0)
-
-    daily_provider_metrics.each do |m|
-      date = m.timestamp.strftime("%Y-%m-%d")
-      fill_pod_metrics(m, date, daily_pod_create_trend, daily_pod_delete_trend)
-    end
-
-    if daily_pod_create_trend.size > 1
-      {
-        :interval_name => "daily",
-        :xy_data       => create_delete_data(daily_pod_create_trend, daily_pod_delete_trend)
-      }
-    end
-  end
-
-  def pod_metrics
-    daily_pod_metrics || hourly_pod_metrics
-  end
-
-  def create_delete_data(create_trend, delete_trend)
-    if create_trend.any?
-      {
-        :xData    => create_trend.keys,
-        :yCreated => create_trend.values.map(&:round),
-        :yDeleted => delete_trend.values.map(&:round)
-      }
-    end
+    daily_network_metrics || hourly_network_metrics || realtime_network_metrics || empty_network_trend_data
   end
 
   def hourly_image_metrics
     hourly_image_metrics = Hash.new(0)
-    hourly_provider_metrics.each do |m|
+    hourly_metrics.each do |m|
       hour = m.timestamp.beginning_of_hour.utc
       if m.stat_container_image_registration_rate.present?
         hourly_image_metrics[hour] += m.stat_container_image_registration_rate
@@ -375,7 +202,7 @@ class ContainerDashboardService
 
   def daily_image_metrics
     daily_image_metrics = Hash.new(0)
-    daily_provider_metrics.each do |m|
+    daily_metrics.each do |m|
       day = m.timestamp.strftime("%Y-%m-%d")
       daily_image_metrics[day] +=
         m.stat_container_image_registration_rate if m.stat_container_image_registration_rate.present?
@@ -391,36 +218,5 @@ class ContainerDashboardService
 
   def image_metrics
     daily_image_metrics || hourly_image_metrics
-  end
-
-  def trend_data(trend)
-    if trend.any?
-      {
-        :xData => trend.keys,
-        :yData => trend.values.map(&:round)
-      }
-    end
-  end
-
-  def realtime_provider_metrics
-    current_user = @controller.current_user
-    tp = TimeProfile.profile_for_user_tz(current_user.id, current_user.get_timezone) || TimeProfile.default_time_profile
-    Metric::Helper.find_for_interval_name('realtime', tp)
-                  .where(:resource => (@ems.try(:container_nodes) || ContainerNode.all))
-                  .where('timestamp > ?', REALTIME_TIME_RANGE.minutes.ago.utc).order('timestamp')
-  end
-
-  def hourly_provider_metrics
-    MetricRollup.with_interval_and_time_range("hourly", (1.day.ago.beginning_of_hour.utc)..(Time.now.utc))
-                .where(:resource => (@ems || ManageIQ::Providers::ContainerManager.all))
-  end
-
-  def daily_provider_metrics
-    current_user = @controller.current_user
-    tp = TimeProfile.profile_for_user_tz(current_user.id, current_user.get_timezone) || TimeProfile.default_time_profile
-
-    @daily_metrics ||= Metric::Helper.find_for_interval_name('daily', tp)
-                                     .where(:resource => (@ems || ManageIQ::Providers::ContainerManager.all))
-                                     .where('timestamp > ?', 30.days.ago.utc).order('timestamp')
   end
 end
